requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==8.0.2
bcrypt==4.3.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
psycopg2-binary